# Conflicts:
#	TMessagesProj/src/main/java/org/telegram/messenger/BuildVars.java

+++ TMessagesProj/src/main/java/org/telegram/ui/Components/AudioPlayerAlert.java
@@ -460,7 +460,7 @@ public class AudioPlayerAlert extends BottomSheet implements NotificationCenter.
             } else {
                 TLRPC.Chat chat = MessagesController.getInstance(currentAccount).getChat(-did);
                 if (chat != null) {
-                    String title = UserConfig.getChatTitleOverride(currentAccount, (int)chat.id);
+                    String title = UserConfig.getChatTitleOverride(currentAccount, chat.id);
                     if (title == null) {
                         title = chat.title;
                     }
